style: create new colors patherns
