+++ index.html
@@ -177,7 +177,7 @@
                 <div class="field">
                     <label for="fieldMessage">Como posso te ajudar?</label>
                     <textarea id="fieldMessage" name="message"
-                        placeholder="Me conte sobre seu projeto, duvida ou mande um oi"></textarea>
+                        placeholder="Me conte sobre seu projeto, duvida ou mande um oi" rows="4"></textarea>
                     <legend></legend>
                 </div>
                 <input type="hidden" name="_captcha" value="false">
@@ -192,7 +192,7 @@
         <section class="container certificates">
             <h2>Minhas certificações</h2>
             <div class="certificate-box">
-                <img src="./public/certificates/FEUX.png">
+                <img src="./public/certificates/FEUX.png" alt="certificado curso de Figma">
                 <details>
                     <summary>Figma by FEUX Felipe Santana
                     </summary>
@@ -203,7 +203,7 @@
                 </details>
             </div>
             <div class="certificate-box">
-                <img src="./public/certificates/kenzie.png">
+                <img src="./public/certificates/kenzie.png" alt="certificado curso programação kenzie academy">
                 <details>
                     <summary> Full Stack Javascript by Kenzie Academy
                     </summary>
@@ -213,7 +213,7 @@
                 </details>
             </div>
             <div class="certificate-box">
-                <img src="./public/certificates/google.png">
+                <img src="./public/certificates/google.png" alt="certificado curso ux designer">
                 <details>
                     <summary> UX Design by Google
                     </summary>
@@ -226,7 +226,7 @@
             </div>
 
             <div class="certificate-box">
-                <img src="./public/certificates/Voitto.png">
+                <img src="./public/certificates/Voitto.png" alt="certificado curso scrum">
                 <details>
                     <summary>Agile Scrum by Grupo Voitto
                     </summary>
@@ -238,7 +238,7 @@
             </div>
 
             <div class="certificate-box">
-                <img src="./public/certificates/Fiap.png">
+                <img src="./public/certificates/Fiap.png" alt="certificado curso de Linux">
                 <details>
                     <summary>Linux by FIAP
                     </summary>
@@ -250,7 +250,7 @@
             </div>
 
             <div class="certificate-box">
-                <img src="./public/certificates/Fiap.png">
+                <img src="./public/certificates/Fiap.png" alt="certificado curso de design thinking">
                 <details>
                     <summary>Design thinking by FIAP
                     </summary>
@@ -262,7 +262,7 @@
             </div>
 
             <div class="certificate-box">
-                <img src="./public/certificates/dio.png">
+                <img src="./public/certificates/dio.png" alt="certificado curso de git e github">
                 <details>
                     <summary>Git e Github by DIO.
                     </summary>
