+++ robots.txt
@@ -1,0 +1,5 @@
+User-agent: Googlebot
+User-agent: *
+Disallow: /nogooglebot/
+Allow: /
+Sitemap: https://dener.criarbr.com/sitemap.xml
