+++ sitemap.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset
+      xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
+            http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
+<!-- created with Free Online Sitemap Generator www.xml-sitemaps.com -->
+
+
+<url>
+  <loc>https://dener.criarbr.com/</loc>
+  <lastmod>2023-08-24T00:29:50+00:00</lastmod>
+</url>
+<url>
+  <loc>https://dener.criarbr.com/blog</loc>
+  <lastmod>2023-08-24T00:29:50+00:00</lastmod>
+</url>
+<url>
+  <loc>https://dener.criarbr.com/pop-cinema</loc>
+  <lastmod>2023-08-24T00:29:50+00:00</lastmod>
+</url>
+<url>
+  <loc>https://dener.criarbr.com/lit-academy</loc>
+  <lastmod>2023-08-24T00:29:50+00:00</lastmod>
+</url>
+<url>
+  <loc>https://dener.criarbr.com/company-geradores</loc>
+  <lastmod>2023-08-24T00:29:50+00:00</lastmod>
+</url>
+
+</urlset>
