+++ src/pages/blog.html
@@ -20,7 +20,7 @@
 
     <link rel="canonical" href="https://dener.criarbr.com/">
     <link rel="sitemap" href="sitemap.xml">
-    <link rel="icon" type="image/svg+xml" href="../public/assets/avatar.ico">
+    <link rel="icon" type="image/svg+xml" href="/public/assets/avatar.ico">
 
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -50,18 +50,19 @@
             <a class="btn-primary" href="/src/pages/blog.html">Blog</a>
         </nav>
         <div class="network">
-            <a href="https://www.linkedin.com/in/denergarcia/" target="_blank" aria-label="Perfil no Linkedin"><i
-                    class="ph ph-linkedin-logo"></i></a>
-            <a href="https://github.com/Dener-Garcia" target="_blank" aria-label="Perfil no Github"> <i
+            <a class="btn-icon" href="https://www.linkedin.com/in/denergarcia/" target="_blank"
+                aria-label="Perfil no Linkedin"><i class="ph ph-linkedin-logo"></i></a>
+            <a class="btn-icon" href="https://github.com/Dener-Garcia" target="_blank" aria-label="Perfil no Github"> <i
                     class="ph ph-github-logo"></i></a>
-            <a href="mailto:dener.criarbr@gmail.com?cc=dener@criarbr.com&subject=Contato via site&body=Ola conheci seu site portfolio e gostaria de..."
+            <a class="btn-icon"
+                href="mailto:dener.criarbr@gmail.com?cc=dener@criarbr.com&subject=Contato via site&body=Ola conheci seu site portfolio e gostaria de..."
                 target="_blank" aria-label="Enviar um e-mail"> <i class="ph ph-envelope-simple"></i></a>
-            <a href="https://www.youtube.com/@dener.criarbr" target="_blank" aria-label="Canal no Youtube"><i
-                    class="ph ph-youtube-logo"></i></a>
-            <a href="https://www.instagram.com/dener.criarbr" target="_blank" aria-label="Perfil no Instagram"><i
-                    class="ph ph-instagram-logo"></i></a>
-            <a href="https://www.behance.net/dener-garcia" target="_blank" aria-label="Perfil no Behance"><i
-                    class="ph ph-behance-logo"></i></a>
+            <a class="btn-icon" href="https://www.youtube.com/@dener.criarbr" target="_blank"
+                aria-label="Canal no Youtube"><i class="ph ph-youtube-logo"></i></a>
+            <a class="btn-icon" href="https://www.instagram.com/dener.criarbr" target="_blank"
+                aria-label="Perfil no Instagram"><i class="ph ph-instagram-logo"></i></a>
+            <a class="btn-icon" href="https://www.behance.net/dener-garcia" target="_blank"
+                aria-label="Perfil no Behance"><i class="ph ph-behance-logo"></i></a>
         </div>
         <blockquote><q cite="https://www.instagram.com/dener.criarbr">People ignore design that ignores people!</q>
         </blockquote>
@@ -99,7 +100,7 @@
     <main>
         <section class="container tutorials">
             <h1>Olá, o que vamos aprender hoje?</h1>
-            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o design até a programação.</p>
             <div class="grid-cards"></div>
         </section>
     </main>
