+++ src/pages/company-geradores.html
@@ -82,7 +82,7 @@
                 <a class="btn-primary" href="https://companygeradores.com/" target="_blank">Visitar Site</a>
                 <a class="btn-primary" href="https://www.figma.com/file/8xJgXalmLL28l2V9e8oo2v/Site-Company-Geradores?node-id=429%3A483&t=ijffjey43WJtehL1-1" target="_blank">Veja no Figma</a>
             </div>
-            <img src="../../public/assets/projects/company-geradores/cover.jpg">
+            <img src="../../public/assets/projects/company-geradores/cover.jpg" alt="Estudo de caso">
 
             <article class="infos">
                 <div class="title-description">
@@ -92,7 +92,7 @@
                         <p>Criar a presença digital da Company Geradores através de um site que seja totalmente responsivo, gerando valor para seus usuários.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/company-geradores/mobile-screens.jpg">
+                <img src="../../public/assets/projects/company-geradores/mobile-screens.jpg" alt="Estudo de caso">
                 <div class="title-description">
                     <h3>Kick off</h3>
                     <div class="description">
@@ -100,7 +100,7 @@
                         <p>O projeto estava definido que seria um site one page, foquei bastante na acessibilidade, constraste de cores e passar informações de forma clara e objetiva fazendo um trabalho forte de UX writing.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/company-geradores/value.jpg">
+                <img src="../../public/assets/projects/company-geradores/value.jpg" alt="Estudo de caso">
 
 
                 <div class="title-description">
@@ -110,8 +110,8 @@
                         <p>Um layout simples mas que tem muita informação a ser passada, estruturei a maioria em forma de cards e tópicos, o destaque maior para os botões de contato e os tipos de serviços que são oferecidos pela empresa.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/company-geradores/Sitemap.jpg">
-                <img src="../../public/assets/projects/company-geradores/Wireframe.jpg">
+                <img src="../../public/assets/projects/company-geradores/Sitemap.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/company-geradores/Wireframe.jpg" alt="Estudo de caso">
 
                 <div class="title-description">
                     <h3>Handoff</h3>
@@ -121,9 +121,9 @@
                         <p>Uma das exigências do projeto era a utilização das cores azul e laranja que tiveram de ser adaptadas para o uso em telas, como foquei em um design clean não tive problemas na hora do desenvolvimento.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/company-geradores/Style-guide.jpg">
-                <img src="../../public/assets/projects/company-geradores/Components.jpg">
-                <img src="../../public/assets/projects/company-geradores/Performance.jpg">
+                <img src="../../public/assets/projects/company-geradores/Style-guide.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/company-geradores/Components.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/company-geradores/Performance.jpg" alt="Estudo de caso">
 
                 <div class="title-description">
                     <h3>Desafios</h3>
