+++ src/pages/lit-academy.html
@@ -82,7 +82,7 @@
             <div class="links">
                 <a class="btn-primary" href="https://www.figma.com/file/kXy0KV1Kd7vSAJ63zmtnnP/Landing-Page-Lit-Academy---Challenge?node-id=2717%3A223&t=C99c96ERvJmTHU4p-1" target="_blank">Veja no Figma</a>
             </div>
-            <img src="../../public/assets/projects/lit-academy/Cover.jpg">
+            <img src="../../public/assets/projects/lit-academy/Cover.jpg" alt="Estudo de caso">
 
             <article class="infos">
                 <div class="title-description">
@@ -92,7 +92,7 @@
                         <p>Criar uma landing page moderna sobre uma empresa de inteligência artificial destacando sessões como apresentação da empresa, produtos oferecidos, como funciona o produto, depoimento de clientes e um formulário para contato.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/lit-academy/screen-desktop.jpg">
+                <img src="../../public/assets/projects/lit-academy/screen-desktop.jpg" alt="Estudo de caso">
                 <div class="title-description">
                     <h3>Kick off</h3>
                     <div class="description">
@@ -100,7 +100,7 @@
                         <p>Com base no moodboard e outras referências de design, busquei construir uma página simples com poucas sessões, mas que apresente o novo produto de uma empresa que trabalhar com IA do tipo interativa para chats.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/lit-academy/Moodboard.jpg">
+                <img src="../../public/assets/projects/lit-academy/Moodboard.jpg" alt="Estudo de caso">
 
 
                 <div class="title-description">
@@ -111,10 +111,10 @@
                         <p>Utilizei um grid bastante comum de 8 pontos e trouxe na hierarquia das informações como imagens, títulos e botões para separar bem as sessões de cada parte da página.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/lit-academy/UserFlow.jpg">
-                <img src="../../public/assets/projects/lit-academy/Wireframe.jpg">
-                <img src="../../public/assets/projects/lit-academy/screen-desktop1.jpg">
-                <img src="../../public/assets/projects/lit-academy/screen-mobile.jpg">
+                <img src="../../public/assets/projects/lit-academy/UserFlow.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/Wireframe.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/screen-desktop1.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/screen-mobile.jpg" alt="Estudo de caso">
 
                 <div class="title-description">
                     <h3>Handoff</h3>
@@ -123,10 +123,10 @@
                         <p>Como designer temos a responsábilidade de pensar em performance também, criar uma interface atraente mas que seja possível de "codar", no fim uma experiência boa de usuário esta ligado a vários outros fatores como resposta de carregamento e interações.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/lit-academy/colors.jpg">
-                <img src="../../public/assets/projects/lit-academy/type.jpg">
-                <img src="../../public/assets/projects/lit-academy/grid.jpg">
-                <img src="../../public/assets/projects/lit-academy/Components.jpg">
+                <img src="../../public/assets/projects/lit-academy/colors.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/type.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/grid.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/lit-academy/Components.jpg" alt="Estudo de caso">
 
                 <div class="title-description">
                     <h3>Desafios</h3>
