+++ src/pages/pop-cinema.html
@@ -82,7 +82,7 @@
             <div class="links">
                 <a class="btn-primary" href="https://www.figma.com/file/vNgFSBkHMSvulrEnvEsaYp/Pop-Cinema---UX-Designe-Google?type=design&node-id=0-1&t=8plJHinYDmHmPtrQ-0" target="_blank">Veja no Figma</a>
             </div>
-            <img src="../../public/assets/projects/pop-cinema/mockups3.jpg">
+            <img src="../../public/assets/projects/pop-cinema/mockups3.jpg" alt="Estudo de caso">
 
             <article class="infos">
                 <div class="title-description">
@@ -92,7 +92,7 @@
                         <p>Criar uma solução com base nos conhecimentos adquiridos durante os seis meses do curso de UX do Google, explorar os meéodos e processos de UX como matriz CSD, entrevistas com usuários, detalhar e entender os dados para construir uma interface que seja adequada, acessível e responsiva para todos os usuários.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/pop-cinema/Cover.jpg">
+                <img src="../../public/assets/projects/pop-cinema/Cover.jpg" alt="Estudo de caso">
                 <div class="title-description">
                     <h3>Kick off</h3>
                     <div class="description">
@@ -100,15 +100,15 @@
                         <p>Após a definição do projeto, parti para a descoberta, fazendo uso de ferramentas como matriz CSD, e frameworks como o Design Thinking a fim de ter bem definido todos os processos pelo qual o projeto necessitaria passar.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/pop-cinema/study-case1.jpg">
-                <img src="../../public/assets/projects/pop-cinema/overview.jpg">
-                <img src="../../public/assets/projects/pop-cinema/overview2.jpg">
-                <img src="../../public/assets/projects/pop-cinema/overview3.jpg">
-                <img src="../../public/assets/projects/pop-cinema/CSD Matrix.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Research.jpg">
-                <img src="../../public/assets/projects/pop-cinema/dot-pain.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Persona-1.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Persona-2.jpg">
+                <img src="../../public/assets/projects/pop-cinema/study-case1.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/overview.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/overview2.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/overview3.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/CSD Matrix.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Research.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/dot-pain.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Persona-1.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Persona-2.jpg" alt="Estudo de caso">
 
 
                 <div class="title-description">
@@ -118,14 +118,14 @@
                         <p>Transformar ideias em soluções, tendo em vista os dados levantados por pesquisas, benchmark e desk research, chegou a hora de estruturar a interface. User flow, user journey e com um site map o proximo passo são os wireframes do nosso aplicativo.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/pop-cinema/user-map.jpg">
-                <img src="../../public/assets/projects/pop-cinema/user-flow.jpg">
-                <img src="../../public/assets/projects/pop-cinema/site-map.jpg">
-                <img src="../../public/assets/projects/pop-cinema/design2.jpg">
-                <img src="../../public/assets/projects/pop-cinema/design3.jpg">
-                <img src="../../public/assets/projects/pop-cinema/design4.jpg">
-                <img src="../../public/assets/projects/pop-cinema/mockups1.jpg">
-                <img src="../../public/assets/projects/pop-cinema/mockups2.jpg">
+                <img src="../../public/assets/projects/pop-cinema/user-map.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/user-flow.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/site-map.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/design2.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/design3.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/design4.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/mockups1.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/mockups2.jpg" alt="Estudo de caso">
 
                 <div class="title-description">
                     <h3>Handoff</h3>
@@ -134,10 +134,10 @@
                         <p>A integração entre design e código nem sempre é fluida já que ambos trabalham ao mesmo tempo no mesmo projeto, ter bem definido o tipo de entrega, componentização do projeto e uma boa comunicação visual é fundamental para a dead line em qualquer projeto de sucesso.</p>
                     </div>
                 </div>
-                <img src="../../public/assets/projects/pop-cinema/Colors.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Typograph.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Components.jpg">
-                <img src="../../public/assets/projects/pop-cinema/Components-1.jpg">
+                <img src="../../public/assets/projects/pop-cinema/Colors.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Typograph.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Components.jpg" alt="Estudo de caso">
+                <img src="../../public/assets/projects/pop-cinema/Components-1.jpg" alt="Estudo de caso">
                 <div class="title-description">
                     <h3>Desafios</h3>
                     <div class="description">
