+++ src/styles/modules/_buttons.scss
@@ -116,3 +116,56 @@
   transform: translateY(0px) rotate(-45deg) scale(1.3);
   background-color: var(--cl-feedback-error);
 }
+
+.show-menu-mobile {
+  display: none;
+
+  @media screen and (max-width: 959px) {
+    display: block;
+    position: fixed;
+    z-index: 10;
+    right: $fs-16;
+    top: $fs-16;
+    width: $fs-40;
+    height: $fs-40;
+    border-radius: $fs-8;
+    background-color: var(--bg-primary);
+    box-shadow: $effect-shadow-default;
+    color: var(--bg-constrast-light);
+    overflow: hidden;
+    cursor: pointer;
+  }
+
+  span:nth-child(1) {
+    position: absolute;
+    width: 16px;
+    height: 4px;
+    left: 10px;
+    top: 10px;
+    border-radius: 5rem;
+    background-color: var(--bg-constrast-light);
+    transition: 0.2s ease-in-out
+  }
+
+  span:nth-child(2) {
+    position: absolute;
+    width: 20px;
+    height: 4px;
+    left: 10px;
+    top: 18px;
+    border-radius: 5rem;
+    background-color: var(--bg-constrast-light);
+    transition: 0.2s ease-in-out
+  }
+
+  span:nth-child(3) {
+    position: absolute;
+    width: 8px;
+    height: 4px;
+    left: 10px;
+    top: 26px;
+    border-radius: 5rem;
+    background-color: var(--bg-constrast-light);
+    transition: 0.2s ease-in-out
+  }
+}
