+++ src/styles/modules/_class-by-js.scss
@@ -50,3 +50,8 @@
 .btnReturnPosition{
     right: $fs-16;
 }
+
+.highLight{
+    box-shadow: 0px 0px 4px #fff;
+    outline: $fs-2 solid #fff;
+}
